requirements.txt
pip
pyfiglet
speedtest-cli
requests
colorama
fast-cli
